# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "edr-isobaric",
      "summary": "EDR API for isobaric data",
      "description": "See <https://api.met.no/weatherapi/isobaricgrib/1.0/documentation>",
      "termsOfService": "https://api.met.no/doc/TermsOfService",
      "contact": {
        "name": "The Norwegian Meteorological Institute",
        "url": "https://api.met.no/doc/support",
        "email": "weatherapi-adm@met.no"
      },
      "license": {
        "name": "Norwegian Licence for Open Government Data (NLOD) 2.0",
        "url": "https://api.met.no/doc/License"
      },
      "version": "0.9.0"
    },
    "paths": {
...
      "/collections/isobaric/position": {
        "get": {
          "tags": [
            "Collection Data"
          ],
          "summary": "Get Isobaric Page",
          "description": "Return data closest to a position.",
          "operationId": "get_isobaric_page_collections_isobaric_position_get",
          "parameters": [
            {
              "name": "coords",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string",
                "minLength": 9,
                "pattern": "^POINT\\(\\d+\\.?\\d* \\d+\\.?\\d*\\)$",
                "description": "Coordinates, formated as a WKT point: POINT(11.9384 60.1699)",
                "title": "Coords"
              },
              "description": "Coordinates, formated as a WKT point: POINT(11.9384 60.1699)",
              "examples": {
                "Oslo": {
                  "summary": "Henrik Mohns Plass 1",
                  "description": "Fetch data for a position in Oslo",
                  "value": "POINT(11.9384 60.1699)"
                },
                "Asker": {
                  "summary": "Asker",
                  "description": "Fetch data for a position in Asker",
                  "value": "POINT(10.45 59.83)"
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Coverage"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Axes": {
          "properties": {
            "x": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ValuesAxis_float_"
                },
                {
                  "$ref": "#/components/schemas/CompactAxis"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X"
            },
            "y": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ValuesAxis_float_"
                },
                {
                  "$ref": "#/components/schemas/CompactAxis"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Y"
            },
            "z": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ValuesAxis_float_"
                },
                {
                  "$ref": "#/components/schemas/CompactAxis"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Z"
            },
            "t": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ValuesAxis_AwareDatetime_"
                },
                {
                  "type": "null"
                }
              ]
            },
            "composite": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ValuesAxis_Tuple_"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          "additionalProperties": false,
          "type": "object",
          "title": "Axes"
        },
...
        "CompactAxis": {
          "properties": {
            "start": {
              "type": "number",
              "title": "Start"
            },
            "stop": {
              "type": "number",
              "title": "Stop"
            },
            "num": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Num"
            }
          },
          "additionalProperties": false,
          "type": "object",
          "required": [
            "start",
            "stop",
            "num"
          ],
          "title": "CompactAxis"
        },
...
        "Coverage": {
          "properties": {
            "id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Id"
            },
            "type": {
              "const": "Coverage",
              "title": "Type",
              "default": "Coverage"
            },
            "domain": {
              "$ref": "#/components/schemas/Domain"
            },
            "parameters": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/covjson_pydantic__parameter__Parameter"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parameters"
            },
            "parameterGroups": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ParameterGroup"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parametergroups"
            },
            "ranges": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/NdArray"
                  },
                  {
                    "$ref": "#/components/schemas/TiledNdArray"
                  },
                  {
                    "type": "string",
                    "minLength": 1,
                    "format": "uri"
                  }
                ]
              },
              "type": "object",
              "title": "Ranges"
            }
          },
          "additionalProperties": true,
          "type": "object",
          "required": [
            "domain",
            "ranges"
          ],
          "title": "Coverage"
        },
...
        "DataType": {
          "type": "string",
          "const": "float",
          "title": "DataType"
        },
        "Domain": {
          "properties": {
            "type": {
              "const": "Domain",
              "title": "Type",
              "default": "Domain"
            },
            "domainType": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DomainType"
                },
                {
                  "type": "null"
                }
              ]
            },
            "axes": {
              "$ref": "#/components/schemas/Axes"
            },
            "referencing": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ReferenceSystemConnectionObject"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Referencing"
            }
          },
          "additionalProperties": true,
          "type": "object",
          "required": [
            "axes"
          ],
          "title": "Domain"
        },
        "DomainType": {
          "type": "string",
          "enum": [
            "Grid",
            "VerticalProfile",
            "PointSeries",
            "Point",
            "MultiPointSeries",
            "MultiPoint"
          ],
          "title": "DomainType"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "NdArray": {
          "properties": {
            "type": {
              "const": "NdArray",
              "title": "Type",
              "default": "NdArray"
            },
            "dataType": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DataType"
                }
              ],
              "default": "float"
            },
            "axisNames": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Axisnames"
            },
            "shape": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shape"
            },
            "values": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array",
              "title": "Values"
            }
          },
          "additionalProperties": true,
          "type": "object",
          "required": [
            "values"
          ],
          "title": "NdArray"
        },
        "ParameterGroup": {
          "properties": {
            "type": {
              "const": "ParameterGroup",
              "title": "Type",
              "default": "ParameterGroup"
            },
            "id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Id"
            },
            "label": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Label"
            },
            "description": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "observedProperty": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/covjson_pydantic__observed_property__ObservedProperty"
                },
                {
                  "type": "null"
                }
              ]
            },
            "members": {
              "items": {
                "type": "string"
              },
              "type": "array",
              "title": "Members"
            }
          },
          "additionalProperties": true,
          "type": "object",
          "required": [
            "members"
          ],
          "title": "ParameterGroup"
        },
...
        "ReferenceSystem": {
          "properties": {
            "type": {
              "type": "string",
              "enum": [
                "GeographicCRS",
                "ProjectedCRS",
                "VerticalCRS",
                "TemporalRS",
                "IdentifierRS"
              ],
              "title": "Type"
            },
            "id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Id"
            },
            "description": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "calendar": {
              "anyOf": [
                {
                  "const": "Gregorian"
                },
                {
                  "type": "string",
                  "minLength": 1,
                  "format": "uri"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Calendar"
            },
            "timeScale": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "format": "uri"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Timescale"
            },
            "label": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Label"
            },
            "targetConcept": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TargetConcept"
                },
                {
                  "type": "null"
                }
              ]
            },
            "identifiers": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/TargetConcept"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Identifiers"
            }
          },
          "additionalProperties": true,
          "type": "object",
          "required": [
            "type"
          ],
          "title": "ReferenceSystem"
        },
        "ReferenceSystemConnectionObject": {
          "properties": {
            "coordinates": {
              "items": {
                "type": "string"
              },
              "type": "array",
              "title": "Coordinates"
            },
            "system": {
              "$ref": "#/components/schemas/ReferenceSystem"
            }
          },
          "additionalProperties": false,
          "type": "object",
          "required": [
            "coordinates",
            "system"
          ],
          "title": "ReferenceSystemConnectionObject"
        },
...
        "TargetConcept": {
          "properties": {
            "id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Id"
            },
            "label": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object",
              "title": "Label"
            },
            "description": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            }
          },
          "additionalProperties": false,
          "type": "object",
          "required": [
            "label"
          ],
          "title": "TargetConcept"
        },
...
        "TileSet": {
          "properties": {
            "tileShape": {
              "items": {
                "anyOf": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array",
              "title": "Tileshape"
            },
            "urlTemplate": {
              "type": "string",
              "title": "Urltemplate"
            }
          },
          "additionalProperties": false,
          "type": "object",
          "required": [
            "tileShape",
            "urlTemplate"
          ],
          "title": "TileSet"
        },
        "TiledNdArray": {
          "properties": {
            "type": {
              "const": "TiledNdArray",
              "title": "Type",
              "default": "TiledNdArray"
            },
            "dataType": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/DataType"
                }
              ],
              "default": "float"
            },
            "axisNames": {
              "items": {
                "type": "string"
              },
              "type": "array",
              "title": "Axisnames"
            },
            "shape": {
              "items": {
                "type": "integer"
              },
              "type": "array",
              "title": "Shape"
            },
            "tileSets": {
              "items": {
                "$ref": "#/components/schemas/TileSet"
              },
              "type": "array",
              "title": "Tilesets"
            }
          },
          "additionalProperties": true,
          "type": "object",
          "required": [
            "axisNames",
            "shape",
            "tileSets"
          ],
          "title": "TiledNdArray"
        },
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        },
        "ValuesAxis_AwareDatetime_": {
          "properties": {
            "dataType": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datatype"
            },
            "coordinates": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Coordinates"
            },
            "values": {
              "items": {
                "type": "string",
                "format": "date-time"
              },
              "type": "array",
              "title": "Values"
            },
            "bounds": {
              "anyOf": [
                {
                  "items": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bounds"
            }
          },
          "additionalProperties": true,
          "type": "object",
          "required": [
            "values"
          ],
          "title": "ValuesAxis[AwareDatetime]"
        },
        "ValuesAxis_Tuple_": {
          "properties": {
            "dataType": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datatype"
            },
            "coordinates": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Coordinates"
            },
            "values": {
              "items": {
                "items": {},
                "type": "array"
              },
              "type": "array",
              "title": "Values"
            },
            "bounds": {
              "anyOf": [
                {
                  "items": {
                    "items": {},
                    "type": "array"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bounds"
            }
          },
          "additionalProperties": true,
          "type": "object",
          "required": [
            "values"
          ],
          "title": "ValuesAxis[Tuple]"
        },
        "ValuesAxis_float_": {
          "properties": {
            "dataType": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Datatype"
            },
            "coordinates": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Coordinates"
            },
            "values": {
              "items": {
                "type": "number"
              },
              "type": "array",
              "title": "Values"
            },
            "bounds": {
              "anyOf": [
                {
                  "items": {
                    "type": "number"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bounds"
            }
          },
          "additionalProperties": true,
          "type": "object",
          "required": [
            "values"
          ],
          "title": "ValuesAxis[float]"
        },
...
        "covjson_pydantic__observed_property__Category": {
          "properties": {
            "id": {
              "type": "string",
              "title": "Id"
            },
            "label": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object",
              "title": "Label"
            },
            "description": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            }
          },
          "additionalProperties": false,
          "type": "object",
          "required": [
            "id",
            "label"
          ],
          "title": "Category"
        },
        "covjson_pydantic__observed_property__ObservedProperty": {
          "properties": {
            "id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Id"
            },
            "label": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object",
              "title": "Label"
            },
            "description": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "categories": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/covjson_pydantic__observed_property__Category"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categories"
            }
          },
          "additionalProperties": false,
          "type": "object",
          "required": [
            "label"
          ],
          "title": "ObservedProperty"
        },
        "covjson_pydantic__parameter__Parameter": {
          "properties": {
            "type": {
              "const": "Parameter",
              "title": "Type",
              "default": "Parameter"
            },
            "id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Id"
            },
            "label": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Label"
            },
            "description": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "observedProperty": {
              "$ref": "#/components/schemas/covjson_pydantic__observed_property__ObservedProperty"
            },
            "categoryEncoding": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "items": {
                          "type": "integer"
                        },
                        "type": "array"
                      }
                    ]
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Categoryencoding"
            },
            "unit": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/covjson_pydantic__unit__Unit"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          "additionalProperties": true,
          "type": "object",
          "required": [
            "observedProperty"
          ],
          "title": "Parameter"
        },
        "covjson_pydantic__unit__Symbol": {
          "properties": {
            "value": {
              "type": "string",
              "title": "Value"
            },
            "type": {
              "type": "string",
              "title": "Type"
            }
          },
          "additionalProperties": false,
          "type": "object",
          "required": [
            "value",
            "type"
          ],
          "title": "Symbol"
        },
        "covjson_pydantic__unit__Unit": {
          "properties": {
            "id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Id"
            },
            "label": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Label"
            },
            "symbol": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "$ref": "#/components/schemas/covjson_pydantic__unit__Symbol"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Symbol"
            }
          },
          "additionalProperties": false,
          "type": "object",
          "title": "Unit"
        },
...
      }
    },
...
  }